Use primitive long diff sets for node state

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/store/record/TokenRecord.java
@@ -28,7 +28,7 @@ public abstract class TokenRecord extends AbstractBaseRecord
     private int nameId;
     private List<DynamicRecord> nameRecords;
 
-    public TokenRecord( int id )
+    public TokenRecord( long id )
     {
         super( id );
     }
